+++ EleFits/EleFits/CompressionStrategy.h
@@ -12,6 +12,19 @@
 namespace Euclid {
 namespace Fits {
 
+/// @cond
+template <long N>
+Position<N> unravel_index(long index, Position<N> shape) // FIXME to Linx
+{
+  for (long i = 0; i < shape.size(); ++i) {
+    const long length = shape[i];
+    shape[i] = index % length;
+    index /= length;
+  }
+  return shape; // FIXME test
+}
+/// @endcond
+
 /**
  * @ingroup compression
  * @brief The interface for implementing compression actions.
@@ -112,16 +125,7 @@ public:
    * then no compression is performed and `false` is returned.
    */
   template <typename T>
-  bool apply(fitsfile* fptr, const ImageHdu::Initializer<T>& init)
-  {
-    // Compress if possible
-    if (auto algo = compression(init)) {
-      Cfitsio::ImageCompression::compress(fptr, *algo);
-      return true;
-    }
-
-    return false;
-  }
+  bool apply(fitsfile* fptr, const ImageHdu::Initializer<T>& init);
 
   /**
    * @brief Try creating a compression algorithm.
@@ -130,25 +134,7 @@ public:
    * then `nullptr` is returned.
    */
   template <typename T>
-  std::unique_ptr<TAlgo> compression(const ImageHdu::Initializer<T>& init)
-  {
-    // CFITSIO does not support 64-bit integer compression
-    if constexpr (bitpix<T>() == 64) {
-      return nullptr;
-    }
-
-    // No compression of data units less than one block long
-    static constexpr std::size_t block_size = 2880;
-    if (shapeSize(init.shape) * sizeof(T) <= block_size) {
-      return nullptr;
-    }
-
-    if (not can_compress(m_algo, init)) {
-      return nullptr;
-    }
-
-    return std::make_unique<TAlgo>(m_algo);
-  }
+  std::unique_ptr<TAlgo> compression(const ImageHdu::Initializer<T>& init);
 
   /// @}
 
@@ -276,61 +262,6 @@ private:
   Compress<TAlgo> m_compress;
 };
 
-/// @cond
-template <typename TAlgo, typename T>
-bool can_compress(const TAlgo&, const ImageHdu::Initializer<T>&)
-{
-  return true;
-}
-
-template <typename T>
-bool can_compress(const Rice& algo, const ImageHdu::Initializer<T>&)
-{
-  return std::is_integral_v<T> || not algo.is_lossless();
-}
-
-template <typename T>
-bool can_compress(const HCompress& algo, const ImageHdu::Initializer<T>& init)
-{
-  const auto& shape = init.shape;
-  if (shapeSize(shape) < 2 || shape[0] < 4 || shape[1] < 4) {
-    return false;
-  }
-  return std::is_integral_v<T> || not algo.is_lossless();
-}
-
-template <typename T>
-bool can_compress(const Plio&, const ImageHdu::Initializer<T>& init)
-{
-  constexpr auto bp = bitpix<T>();
-
-  // Float or too large int
-  if constexpr (bp < 0 || bp > 32) {
-    return false;
-  }
-
-  // Maybe
-  if constexpr (bp > 16) {
-    // Max from records
-    if (init.records.has("DATAMAX")) {
-      const auto max = init.records.template as<T>("DATAMAX");
-      return max < (T(1) << 24);
-    }
-
-    // No max
-    if (not init.data) {
-      return false;
-    }
-
-    // Max from data
-    const auto max = *std::max_element(init.data, init.data + shapeSize(init.shape));
-    return max < (T(1) << 24);
-  }
-
-  return true;
-}
-/// @endcond
-
 /**
  * @ingroup compression
  * @brief A basic adaptive compression strategy.
@@ -395,18 +326,6 @@ private:
    */
   template <typename T>
   Quantization quantization() const;
-
-  /**
-   * @brief Adapt the tiling to the raster shape.
-   */
-  template <typename T>
-  Position<-1> tiling(const ImageHdu::Initializer<T>& init) const;
-
-  /**
-   * @brief Adapt the H-compress tiling to the raster shape.
-   */
-  template <typename T>
-  Position<-1> hcompress_tiling(const ImageHdu::Initializer<T>& init) const;
 
   /**
    * @brief Adapt the H-compress scaling to the raster shape.