+++ EleFits/tests/src/CompressionStrategy_test.cpp
@@ -16,6 +16,80 @@ BOOST_AUTO_TEST_SUITE(CompressionStrategy_test)
 
 //-----------------------------------------------------------------------------
 
+void check_index_to_position(long index, const Position<-1>& shape, const Position<-1>& expected)
+{
+  auto position = unravel_index(index, shape);
+  BOOST_TEST(position == expected);
+}
+
+BOOST_AUTO_TEST_CASE(index_to_position_1d_test)
+{
+  for (long i = 0; i <= 1024; ++i) { // Including out-of-bounds
+    check_index_to_position(i, {1024}, {i});
+  }
+}
+
+BOOST_AUTO_TEST_CASE(index_to_position_2d_test)
+{
+  check_index_to_position(0, {1024, 1024}, {0, 0});
+  check_index_to_position(1023, {1024, 1024}, {1023, 0});
+  check_index_to_position(1024, {1024, 1024}, {0, 1});
+  check_index_to_position(1024 * 1024, {1024, 1024}, {0, 1024});
+}
+
+template <typename T>
+void check_adaptive_tiling(const Position<-1>& shape)
+{
+  Gzip algo;
+  ImageHdu::Initializer<T> init {1, "", {}, shape, nullptr};
+  adapt_tiling(algo, init);
+  if (shapeSize(shape) * sizeof(T) <= 1024 * 1024) {
+    BOOST_TEST((algo.tiling() == shape || algo.tiling() == Tile::whole()));
+  } else {
+    BOOST_TEST(shapeSize(algo.tiling()) * sizeof(T) >= 1024 * 1024);
+    long dim = 6;
+    for (std::size_t i = 0; i < shape.size(); ++i) {
+      const auto length = algo.tiling()[i];
+      if (i < dim && length != shape[i]) {
+        dim = i;
+      }
+      if (i < dim) {
+        BOOST_TEST(length == shape[i]);
+      }
+      if (i == dim) {
+        BOOST_TEST(length <= shape[i]);
+      }
+      if (i > dim) {
+        BOOST_TEST(length == 1);
+      }
+    }
+  }
+}
+
+template <typename T>
+void check_adaptive_tiling()
+{
+  check_adaptive_tiling<T>({});
+  check_adaptive_tiling<T>({1});
+  check_adaptive_tiling<T>({2, 2, 2, 2, 2, 2});
+  check_adaptive_tiling<T>({1024 * 1024});
+  check_adaptive_tiling<T>({1024 * 1024 + 1});
+  check_adaptive_tiling<T>({1, 1024 * 1024});
+  check_adaptive_tiling<T>({1, 1024 * 1024 + 1});
+  check_adaptive_tiling<T>({1, 1024, 1024});
+  check_adaptive_tiling<T>({1, 1024, 1024 + 1});
+  check_adaptive_tiling<T>({1024, 1024, 1024});
+  check_adaptive_tiling<T>({1024, 1024, 1024 + 1});
+}
+
+BOOST_AUTO_TEST_CASE(adaptive_tiling_test)
+{
+  check_adaptive_tiling<unsigned char>();
+  check_adaptive_tiling<int>();
+  check_adaptive_tiling<float>();
+  check_adaptive_tiling<double>();
+}
+
 /**
  * @brief Whatever the type and shape, check losslessness.
  */